+++ apps/cli/design-sync.config.js
@@ -14,12 +14,20 @@ const breakpoints = {
   xxl: '{kda.foundation.breakpoint.xxl}',
 };
 const breakpointKeys = Object.keys(breakpoints);
+const fontOverride = () => "'Haas Grotesk Display', -apple-system, sans-serif";
 export default defineConfig({
+  // uri: 'github:salamaashoush/kda-design-system/tokens#dedupe-tokens',
   uri: 'github:kadena-community/design-system/tokens#main',
+  muteWarnings: true,
   out: 'generated',
   defaultMode: 'light',
   requiredModes: ['light', 'dark'],
   prettify: true,
+  overrides: {
+    'kda.foundation.typography.family.bodyFont': fontOverride,
+    'kda.foundation.typography.family.primaryFont': fontOverride,
+    'kda.foundation.typography.family.headingFont': fontOverride,
+  },
   plugins: [cssPlugin(), vanillaExtractPlugin(), jsonPlugin()],
   schemaExtensions: [
     responsiveExtension({