+++ packages/manager/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @design-sync/manager
 
+## 0.5.0
+
+### Minor Changes
+
+- Improve validation errors/warnnings and implement token overrides
+
+### Patch Changes
+
+- Updated dependencies
+  - @design-sync/w3c-dtfm@0.8.0
+
 ## 0.4.1
 
 ### Patch Changes